+++ src/estimagic/optimization/pounders_history.py
