+++ tests/optimization/test_history.py
@@ -1,7 +1,7 @@
 """Test the history class for least-squares optimizers."""
 import numpy as np
 import pytest
-from estimagic.optimization.history import LeastSquaresHistory
+from estimagic.optimization.pounders_history import LeastSquaresHistory
 from numpy.testing import assert_array_almost_equal as aaae
 
 ENTRIES = [
